Project: REFACTOR ->
moved function specific for AVR ucontroller to seperate files
