+++ test/lcd_hd44780/avr_code_test/CMakeLists.txt
@@ -45,6 +45,7 @@ set(TEST_INCLUDE_DIRS
 
 set(SRCS
 	../../../src/lcd_hd44780.c
+	../../../src/lcd_hd44780_avr_specific.c
 )
 
 set(TEST_SRCS
